Split Up Search Component
